state delete confirmation prompt amd error handler

+++ routes/states/controller.js
@@ -53,10 +53,17 @@ exports.create = async (req, res, next) => {
 exports.removeItem = async (req, res, next) => {
     try {
         let { stateId } = req.params;
+        let {allowDelete} = req.body;
         if (!stateId) throw new Error('Invalide data supplide!')
         let cities = await citymaster.countDocuments({'stateId': stateId})
         if (cities) {
             throw new Error("State can't be removed as City Exist!")
+        }
+        if (!allowDelete) {
+            return res.status(200).json({
+                IsSuccess: true,
+                allowDelete: true,
+            })
         }
         let states = await StateMaster.findByIdAndRemove(stateId);
